+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 * :support:`-` Update ``sdss-drift`` to ``0.1.5`` to fix ``pymodbus`` import.
 * :support:`90` Remove the option to initialise datums in `.Positioner`.
 * :support`-` Wrap message from ``CANNetBus`` in custom message class with additional ``__slots__``. This fixes a deprecation introduced in ``python-can>=3.3.4``.
+* :support:`-` ``motor_steps`` and ``time_step`` are now defined in the configuration file under ``positioner``.
 
 * :release:`0.5.2 <2020-07-31>`
 * :support:`-` Adapt actor system to use ``CLU>=0.3.0``.